[SCons] Fix nondeterminism in source generation
In Fedora rebuilds to test package build reproducibility
(see https://reproducible-builds.org/,
https://fedoraproject.org/wiki/Changes/ReproduciblePackageBuilds),
we get the following differences like the following for various CMakeLists.txt
and SConstruct files:

/usr/share/cantera/samples/cxx/LiC6_electrode/CMakeLists.txt:
│ │ -include_directories("/usr/include" "/usr/include/eigen3" "/usr/include/highfive")
│ │ +include_directories("/usr/include/highfive" "/usr/include/eigen3" "/usr/include")

Those end up in the -debugsources package and cause the whole build to be
flagged as irreproducible. In addition, if the an include file with the same
name happened to be present in more than one location, the unpredictable sort
order would mean that different files would be used in different builds. Sort
the directories alphabetically for predictable results.